+++ lib/app/models/user_settings.dart
@@ -63,6 +63,7 @@ class DefaultUserSettings {
   static const globalProgressEndLabel = ProgressEndLabel.total;
   static const progressEndLabel = ProgressEndLabel.total;
   static const miniplayerSubtitleMode = MiniplayerSubtitleMode.both;
+  static const homeShelves = ShelfIdentity.values;
 }
 
 @freezed
@@ -203,6 +204,8 @@ sealed class UserSettings with _$UserSettings {
 
     @Default(DefaultUserSettings.miniplayerSubtitleMode)
     MiniplayerSubtitleMode miniplayerSubtitleMode,
+
+    @Default(DefaultUserSettings.homeShelves) List<ShelfIdentity> homeShelves,
   }) = _UserSettings;
 
   factory fromJson(Map<String, dynamic> json) => _$UserSettingsFromJson(json);