+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/2.0.0/).
 - Android Auto search
 - OS notification stop button
 - OS notification playback speed cycle
+- reorder and show/hide home screen shelves
 
 ## [v0.7.3] - 2026-09-01
 
@@ -675,6 +676,3 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/2.0.0/).
 [v0.1.2]: https://github.com/likhithpraveenk/storii/compare/v0.1.1...v0.1.2
 [v0.1.1]: https://github.com/likhithpraveenk/storii/compare/v0.1...v0.1.1
 [v0.1]: https://github.com/likhithpraveenk/storii/compare/3fd448b...v0.1
-
-[Unreleased]: compare/v0.7.3...HEAD
-[v0.8.0]: compare/v0.7.3...v0.8.0